# internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for managing HydraDNS server configuration and filtering.",
        "title": "HydraDNS Management API",
        "contact": {
            "name": "HydraDNS Support",
            "url": "https://github.com/jroosing/hydradns"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/filtering/blacklist": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns all domains in the blacklist",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "filtering"
                ],
                "summary": "Get blacklist domains",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.DomainListResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Adds one or more domains to the blacklist",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "filtering"
                ],
                "summary": "Add domains to blacklist",
                "parameters": [
                    {
                        "description": "Domains to add",
                        "name": "domains",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.DomainRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.StatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Removes one or more domains from the blacklist",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "filtering"
                ],
                "summary": "Remove domains from blacklist",
                "parameters": [
                    {
                        "description": "Domains to remove",
                        "name": "domains",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.DomainDeleteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.StatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.ErrorResponse"
                        }
                    },
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/filtering/enabled": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Toggles the filtering engine on or off",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "filtering"
                ],
                "summary": "Enable or disable filtering",
                "parameters": [
                    {
                        "description": "Enable state",
                        "name": "enabled",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.FilteringEnabledRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.StatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/filtering/stats": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns detailed filtering statistics",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "filtering"
                ],
                "summary": "Get filtering statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.FilteringStatsResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/filtering/whitelist": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns all domains in the whitelist",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "filtering"
                ],
                "summary": "Get whitelist domains",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.DomainListResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Adds one or more domains to the whitelist",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "filtering"
                ],
                "summary": "Add domains to whitelist",
                "parameters": [
                    {
                        "description": "Domains to add",
                        "name": "domains",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.DomainRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.StatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Removes one or more domains from the whitelist",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "filtering"
                ],
                "summary": "Remove domains from whitelist",
                "parameters": [
                    {
                        "description": "Domains to remove",
                        "name": "domains",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.DomainDeleteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.StatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.ErrorResponse"
                        }
                    },
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_jroosing_hydradns_internal_api_models.DomainDeleteRequest": {
            "type": "object",
            "required": [
                "domains"
            ],
            "properties": {
                "domains": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "github_com_jroosing_hydradns_internal_api_models.DomainListResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "domains": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "github_com_jroosing_hydradns_internal_api_models.DomainRequest": {
            "type": "object",
            "required": [
                "domains"
            ],
            "properties": {
                "domains": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "github_com_jroosing_hydradns_internal_api_models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "github_com_jroosing_hydradns_internal_api_models.FilteringEnabledRequest": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                }
            }
        },
        "github_com_jroosing_hydradns_internal_api_models.FilteringStatsResponse": {
            "type": "object",
            "properties": {
                "blacklist_size": {
                    "type": "integer"
                },
                "enabled": {
                    "type": "boolean"
                },
                "queries_allowed": {
                    "type": "integer"
                },
                "queries_blocked": {
                    "type": "integer"
                },
                "queries_total": {
                    "type": "integer"
                },
                "whitelist_size": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_jroosing_hydradns_internal_api_models.StatusResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}